move fuselage model to separate file
